requirements.txt
pydantic-ai
docker
flask
python-dotenv
logfire
google-generativeai
